requirements.txt
appdirs==1.4.4
appnope==0.1.0
astroid==2.4.2
backcall==0.2.0
black==20.8b1
click==7.1.2
decorator==4.4.2
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyaml==20.4.0
Pygments==2.7.2
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
wrapt==1.12.1